Make sure setup.py could run without six
Move Patroni class and main function to the __main__.py.
The __init__.py will have only a few functions used by Patroni and from
the setup.py
